V0.7 (#15)
* start impl new particle container

* feat(mc): Change all Monte-Carlo structure using SoA, almost working

* feat(mc): migrate to SoA, working kernel,division,sort and export

* fix: particle_container and divison handling add two_meta model,

fix: remove default constructor, add kokkos_inline for contributions

* fix(exporter): inverse spatial_value export extent order

* feat(events): enable/disable counting event at compile time

* feat(serde): working serde on gpu, adapt unit tests

* chore: clean repo, remove old mc includes

* chore(build): improve model loader generation using a manual list of models

* feat(runtime environement): move kokkos/mpi initialization to Api, old api constructor used in python api depreacted

* chore(model): rename movels and move udfmodel (not working anymore)

* fix(move): __find_next_compartment  use correct probability, add conditional template to kernel

* fix/refractor: refractor and minor fix

* fix(api): adapt python api to new runtime init, remove mpi from python module, init from c++

* fix(api): use new api constructor

* feat(pp): add mass export

* fix(kernel,model): invert indices, and uptake for two_meta model

* documentation: add css doxygen

* refractor(api): remove deprecated

* refractor: remove containerstate, use 2 neighbors view alias for const and non const

* chore: delete biomodels folder

* doc: improve documentation

* minor changes

* documentation improvement

* ci(documentation): improve documentation and ci gh

* fix(api): not init kokkos twice and update unit test

* refractor: minor fix +  compiler warnings

* test(mc): main feature particles_container tested

---------

+++ apps/libs/cma_utils/transitionner/transport.cpp
@@ -2,6 +2,7 @@
 #include <Kokkos_Core.hpp>
 #include <cma_utils/cache_hydro_state.hpp>
 #include <cmt_common/zip.hpp>
+#include <cstddef>
 #include <transitionner/transport.hpp>
 namespace CmaUtils
 {
@@ -30,7 +31,7 @@ namespace CmaUtils
 
     typedef Eigen::Triplet<double> T;
     std::vector<T> tripletList;
-    tripletList.reserve(n_compartments * n_compartments); // Reserve space for all elements
+    tripletList.reserve(static_cast<long long>(n_compartments * n_compartments)); // Reserve space for all elements
 
     // Temporary vector to keep track of the row sums for diagonal elements
     std::vector<double> rowSums(n_compartments, 0.0);
